+++ discord_bots/commands.py
@@ -1,14 +1,12 @@
 import asyncio
 import heapq
 import logging
-from bisect import bisect
-from collections import defaultdict
 from datetime import datetime, timedelta, timezone
 from itertools import combinations
 from math import floor
 from random import choice, shuffle, uniform
 from tempfile import NamedTemporaryFile
-from typing import List, Literal, Optional, Union
+from typing import List, Literal, Optional
 
 import discord
 import imgkit
@@ -32,8 +30,6 @@ from PIL import Image
 from sqlalchemy import func, or_
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm.session import Session as SQLAlchemySession
-from sqlalchemy.sql import select
-from table2ascii import Alignment, PresetStyle, table2ascii
 from trueskill import Rating
 
 import discord_bots.config as config
@@ -41,13 +37,11 @@ from discord_bots.checks import is_admin
 from discord_bots.utils import (
     MU_LOWER_UNICODE,
     SIGMA_LOWER_UNICODE,
-    code_block,
     create_finished_game_embed,
     create_in_progress_game_embed,
     get_team_name_diff,
     mean,
     move_game_players,
-    print_leaderboard,
     send_in_guild_message,
     send_message,
     short_uuid,
@@ -182,112 +176,6 @@ def get_even_teams(
             f"Found team evenness: {best_team_evenness_so_far} interations: {i}"
         )  # DEBUG, TRACE?
         return best_teams_so_far, best_win_prob_so_far
-
-
-# Return n of the most even or least even teams
-# For best teams, use direction = 1, for worst teams use direction = -1
-def get_n_teams(
-    players: list[Player],
-    team_size: int,
-    is_rated: bool,
-    n: int,
-    direction: int = 1,
-) -> list[tuple[list[Player], float]]:
-    teams: list[tuple[float, list[Player]]] = []
-
-    all_combinations = list(combinations(players, team_size))
-    for team0 in all_combinations:
-        team1 = [p for p in players if p not in team0]
-        team0_ratings = list(
-            map(
-                lambda x: Rating(x.rated_trueskill_mu, x.rated_trueskill_sigma),
-                team0,
-            )
-        )
-        team1_ratings = list(
-            map(
-                lambda x: Rating(x.rated_trueskill_mu, x.rated_trueskill_sigma),
-                team1,
-            )
-        )
-        win_prob = win_probability(team0_ratings, team1_ratings)
-        current_team_evenness = abs(0.50 - win_prob)
-        heapq.heappush(
-            teams, (direction * current_team_evenness, list(team0[:]) + list(team1[:]))
-        )
-
-    teams_out = []
-    for _ in range(n):
-        teams_out.append(heapq.heappop(teams))
-
-    return teams_out
-
-
-def get_n_best_teams(
-    players: list[Player], team_size: int, is_rated: bool, n: int
-) -> list[tuple[list[Player], float]]:
-    return get_n_teams(players, team_size, is_rated, n, 1)
-
-
-def get_n_worst_teams(
-    players: list[Player], team_size: int, is_rated: bool, n: int
-) -> list[tuple[list[Player], float]]:
-    return get_n_teams(players, team_size, is_rated, n, -1)
-
-
-# Return n of the most even or least even teams
-# For best teams, use direction = 1, for worst teams use direction = -1
-def get_n_finished_game_teams(
-    fgps: list[FinishedGamePlayer],
-    team_size: int,
-    is_rated: bool,
-    n: int,
-    direction: int = 1,
-) -> list[tuple[list[FinishedGamePlayer], float]]:
-    teams: list[tuple[float, list[FinishedGamePlayer]]] = []
-
-    all_combinations = list(combinations(fgps, team_size))
-    for team0 in all_combinations:
-        team1 = [p for p in fgps if p not in team0]
-        team0_ratings = list(
-            map(
-                lambda x: Rating(
-                    x.rated_trueskill_mu_before, x.rated_trueskill_sigma_before
-                ),
-                team0,
-            )
-        )
-        team1_ratings = list(
-            map(
-                lambda x: Rating(
-                    x.rated_trueskill_mu_before, x.rated_trueskill_sigma_before
-                ),
-                team1,
-            )
-        )
-        win_prob = win_probability(team0_ratings, team1_ratings)
-        current_team_evenness = abs(0.50 - win_prob)
-        heapq.heappush(
-            teams, (direction * current_team_evenness, list(team0[:]) + list(team1[:]))
-        )
-
-    teams_out = []
-    for _ in range(n):
-        teams_out.append(heapq.heappop(teams))
-
-    return teams_out
-
-
-def get_n_best_finished_game_teams(
-    fgps: list[FinishedGamePlayer], team_size: int, is_rated: bool, n: int
-) -> list[tuple[list[FinishedGamePlayer], float]]:
-    return get_n_finished_game_teams(fgps, team_size, is_rated, n, 1)
-
-
-def get_n_worst_finished_game_teams(
-    fgps: list[FinishedGamePlayer], team_size: int, is_rated: bool, n: int
-) -> list[tuple[list[FinishedGamePlayer], float]]:
-    return get_n_finished_game_teams(fgps, team_size, is_rated, n, -1)
 
 
 async def create_game(
@@ -662,200 +550,6 @@ async def add_player_to_queue(
         session.commit()
 
         return True, False
-
-
-def mock_teams_str(
-    team0_players: list[Player],
-    team1_players: list[Player],
-    is_rated: bool,
-) -> str:
-    """
-    Helper method to debug print teams if these were the players
-    """
-    output = ""
-    team0_rating = [
-        Rating(p.rated_trueskill_mu, p.rated_trueskill_sigma) for p in team0_players
-    ]
-    team1_rating = [
-        Rating(p.rated_trueskill_mu, p.rated_trueskill_sigma) for p in team1_players
-    ]
-    team0_names = ", ".join(
-        sorted([escape_markdown(player.name) for player in team0_players])
-    )
-    team1_names = ", ".join(
-        sorted([escape_markdown(player.name) for player in team1_players])
-    )
-    team0_win_prob = round(100 * win_probability(team0_rating, team1_rating), 1)
-    team1_win_prob = round(100 - team0_win_prob, 1)
-    team0_mu = round(mean([player.rated_trueskill_mu for player in team0_players]), 2)
-    team1_mu = round(mean([player.rated_trueskill_mu for player in team1_players]), 2)
-    team0_sigma = round(
-        mean([player.rated_trueskill_sigma for player in team0_players]),
-        2,
-    )
-    team1_sigma = round(
-        mean([player.rated_trueskill_sigma for player in team1_players]),
-        2,
-    )
-    output += f"\n**BE** (**{team0_win_prob}%**, mu: {team0_mu}, sigma: {team0_sigma}): {team0_names}"
-    output += f"\n**DS** (**{team1_win_prob}%**, mu: {team1_mu}, sigma: {team1_sigma}): {team1_names}"
-    return output
-
-
-def mock_finished_game_teams_str(
-    team0_fg_players: list[FinishedGamePlayer],
-    team1_fg_players: list[FinishedGamePlayer],
-    is_rated: bool,
-) -> str:
-    """
-    Helper method to debug print teams if these were the players
-    """
-    output = ""
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        team0_rating = [
-            Rating(fgp.rated_trueskill_mu_before, fgp.rated_trueskill_sigma_before)
-            for fgp in team0_fg_players
-        ]
-        team1_rating = [
-            Rating(fgp.rated_trueskill_mu_before, fgp.rated_trueskill_sigma_before)
-            for fgp in team1_fg_players
-        ]
-        team0_player_ids_map = {x.player_id: x for x in team0_fg_players}
-        team1_player_ids_map = {x.player_id: x for x in team1_fg_players}
-        team0_player_ids = set(map(lambda x: x.player_id, team0_fg_players))
-        team1_player_ids = set(map(lambda x: x.player_id, team1_fg_players))
-        team0_players: list[Player] = session.query(Player).filter(
-            Player.id.in_(team0_player_ids)
-        )
-        team1_players: list[Player] = session.query(Player).filter(
-            Player.id.in_(team1_player_ids)
-        )
-        team0_names = ", ".join(
-            sorted(
-                [
-                    f"{escape_markdown(player.name)} ({round(team0_player_ids_map.get(player.id).rated_trueskill_mu_before, 1)})"
-                    for player in team0_players
-                ]
-            )
-        )
-        team1_names = ", ".join(
-            sorted(
-                [
-                    f"{escape_markdown(player.name)} ({round(team1_player_ids_map.get(player.id).rated_trueskill_mu_before, 1)})"
-                    for player in team1_players
-                ]
-            )
-        )
-        team0_win_prob = round(100 * win_probability(team0_rating, team1_rating), 1)
-        team1_win_prob = round(100 - team0_win_prob, 1)
-        team0_mu = round(
-            mean([player.rated_trueskill_mu_before for player in team0_fg_players]), 2
-        )
-        team1_mu = round(
-            mean([player.rated_trueskill_mu_before for player in team1_fg_players]), 2
-        )
-        team0_sigma = round(
-            mean([player.rated_trueskill_sigma_before for player in team0_fg_players]),
-            2,
-        )
-        team1_sigma = round(
-            mean([player.rated_trueskill_sigma_before for player in team1_fg_players]),
-            2,
-        )
-        output += f"\n**BE** (**{team0_win_prob}%**, mu: {team0_mu}, sigma: {team0_sigma}): {team0_names}"
-        output += f"\n**DS** (**{team1_win_prob}%**, mu: {team1_mu}, sigma: {team1_sigma}): {team1_names}"
-        return output
-
-
-def in_progress_game_str(in_progress_game: InProgressGame, debug: bool = False) -> str:
-    """
-    Helper method to pretty print a finished game
-    """
-    output = ""
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        short_game_id = short_uuid(in_progress_game.id)
-        queue: Queue = (
-            session.query(Queue).filter(Queue.id == in_progress_game.queue_id).first()
-        )
-        if debug:
-            output += f"**{queue.name}** ({short_game_id}) (TS: {round(in_progress_game.average_trueskill, 2)})"
-        else:
-            output += f"**{queue.name}** ({short_game_id})"
-        team0_igp_players: list[InProgressGamePlayer] = session.query(
-            InProgressGamePlayer
-        ).filter(
-            InProgressGamePlayer.in_progress_game_id == in_progress_game.id,
-            InProgressGamePlayer.team == 0,
-        )
-        team1_igp_players: list[InProgressGamePlayer] = session.query(
-            InProgressGamePlayer
-        ).filter(
-            InProgressGamePlayer.in_progress_game_id == in_progress_game.id,
-            InProgressGamePlayer.team == 1,
-        )
-        team0_player_ids = set(map(lambda x: x.player_id, team0_igp_players))
-        team1_player_ids = set(map(lambda x: x.player_id, team1_igp_players))
-        team0_fgp_by_id = {fgp.player_id: fgp for fgp in team0_igp_players}
-        team1_fgp_by_id = {fgp.player_id: fgp for fgp in team1_igp_players}
-        team0_players: list[Player] = session.query(Player).filter(Player.id.in_(team0_player_ids))  # type: ignore
-        team1_players: list[Player] = session.query(Player).filter(Player.id.in_(team1_player_ids))  # type: ignore
-        if debug and False:
-            team0_names = ", ".join(
-                sorted(
-                    [
-                        f"{escape_markdown(player.name)} ({round(team0_fgp_by_id[player.id].rated_trueskill_mu_before, 1)})"
-                        for player in team0_players
-                    ]
-                )
-            )
-            team1_names = ", ".join(
-                sorted(
-                    [
-                        f"{escape_markdown(player.name)} ({round(team1_fgp_by_id[player.id].rated_trueskill_mu_before, 1)})"
-                        for player in team1_players
-                    ]
-                )
-            )
-        else:
-            team0_names = ", ".join(
-                sorted([escape_markdown(player.name) for player in team0_players])
-            )
-            team1_names = ", ".join(
-                sorted([escape_markdown(player.name) for player in team1_players])
-            )
-        # TODO: Include win prob
-        # team0_win_prob = round(100 * finished_game.win_probability, 1)
-        # team1_win_prob = round(100 - team0_win_prob, 1)
-        team0_tsr = round(
-            mean([player.rated_trueskill_mu for player in team0_players]), 1
-        )
-        team1_tsr = round(
-            mean([player.rated_trueskill_mu for player in team1_players]), 1
-        )
-        # TODO: Include win prob
-        if debug:
-            team0_str = f"{in_progress_game.team0_name} ({team0_tsr}): {team0_names}"
-            team1_str = f"{in_progress_game.team1_name} ({team1_tsr}): {team1_names}"
-        else:
-            team0_str = f"{in_progress_game.team0_name} ({team0_names}"
-            team1_str = f"{in_progress_game.team1_name} ({team1_names}"
-
-        output += f"\n{team0_str}"
-        output += f"\n{team1_str}"
-        delta: timedelta = datetime.now(
-            timezone.utc
-        ) - in_progress_game.created_at.replace(tzinfo=timezone.utc)
-        if delta.days > 0:
-            output += f"\n@ {delta.days} days ago\n"
-        elif delta.seconds > 3600:
-            hours_ago = delta.seconds // 3600
-            output += f"\n@ {hours_ago} hours ago\n"
-        else:
-            minutes_ago = delta.seconds // 60
-            output += f"\n@ {minutes_ago} minutes ago\n"
-        return output
 
 
 # Commands start here
@@ -1280,157 +974,6 @@ async def autosub(ctx: Context, member: Member = None):
     session.commit()
 
 
-@bot.command()
-async def commend(ctx: Context, member: Member):
-    session = ctx.session
-    commender: Player | None = (
-        session.query(Player).filter(Player.id == ctx.message.author.id).first()
-    )
-    commendee: Player | None = (
-        session.query(Player).filter(Player.id == member.id).first()
-    )
-    emoji = choice(["🔨", "💥", "🤕", "🤌"])
-    if ctx.message.author.id == member.id:
-        await send_message(
-            ctx.message.channel,
-            embed_description=f"{emoji}  **BONK**  {emoji}",
-            colour=Colour.red(),
-        )
-        return
-
-    if not commendee:
-        await send_message(
-            ctx.message.channel,
-            embed_description=f"Could not find {escape_markdown(member.name)}",
-            colour=Colour.red(),
-        )
-        return
-
-    last_finished_game: FinishedGame | None = (
-        session.query(FinishedGame)
-        .join(FinishedGamePlayer)
-        .filter(FinishedGamePlayer.player_id == commender.id)
-        .order_by(FinishedGame.finished_at.desc())
-        .first()
-    )
-    if not last_finished_game:
-        await send_message(
-            ctx.message.channel,
-            embed_description=f"Could not find last game played for {escape_markdown(member.name)}",
-            colour=Colour.red(),
-        )
-        return
-
-    has_commend = (
-        session.query(Commend)
-        .filter(
-            Commend.finished_game_id == last_finished_game.id,
-            Commend.commender_id == commender.id,
-        )
-        .first()
-    )
-    if has_commend is not None:
-        await send_message(
-            ctx.message.channel,
-            embed_description=f"You already commended someone for this game",
-            colour=Colour.red(),
-        )
-        return
-
-    players_in_last_game = (
-        session.query(FinishedGamePlayer)
-        .filter(FinishedGamePlayer.finished_game_id == last_finished_game.id)
-        .all()
-    )
-    player_ids = set(map(lambda x: x.player_id, players_in_last_game))
-    if commendee.id not in player_ids:
-        await send_message(
-            ctx.message.channel,
-            embed_description=f"{escape_markdown(commendee.name)} was not in your last game",
-            colour=Colour.red(),
-        )
-        return
-
-    session.add(
-        Commend(
-            last_finished_game.id,
-            commender.id,
-            commender.name,
-            commendee.id,
-            commendee.name,
-        )
-    )
-    commender.raffle_tickets += 1
-    session.add(commender)
-    session.commit()
-    await send_message(
-        ctx.message.channel,
-        embed_description=f"⭐ {escape_markdown(commendee.name)} received a commend! ⭐",
-        colour=Colour.green(),
-    )
-    session.close()
-
-
-@bot.command()
-async def commendstats(ctx: Context):
-    session = ctx.session
-    most_commends_given_statement = (
-        select(Player, func.count(Commend.commender_id).label("commend_count"))
-        .join(Commend, Commend.commender_id == Player.id)
-        .group_by(Player.id)
-        .having(func.count(Commend.commender_id) > 0)
-        .order_by(func.count(Commend.commender_id).desc())
-    )
-    most_commends_received_statement = (
-        select(Player, func.count(Commend.commendee_id).label("commend_count"))
-        .join(Commend, Commend.commendee_id == Player.id)
-        .group_by(Player.id)
-        .having(func.count(Commend.commendee_id) > 0)
-        .order_by(func.count(Commend.commendee_id).desc())
-    )
-
-    most_commends_given: List[Player] = session.execute(
-        most_commends_given_statement
-    ).fetchall()
-    most_commends_received: List[Player] = session.execute(
-        most_commends_received_statement
-    ).fetchall()
-    session.close()
-
-    output = "**Most commends given**"
-    for i, row in enumerate(most_commends_given, 1):
-        player = row[Player]
-        count = row["commend_count"]
-        output += f"\n{i}. {count} - {player.name}"
-    output += "\n**Most commends received**"
-    for i, row in enumerate(most_commends_received, 1):
-        player = row[Player]
-        count = row["commend_count"]
-        output += f"\n{i}. {count} - {player.name}"
-
-    if config.LEADERBOARD_CHANNEL:
-        channel = bot.get_channel(config.LEADERBOARD_CHANNEL)
-        await send_message(channel, embed_description=output, colour=Colour.blue())
-        await send_message(
-            ctx.message.channel,
-            embed_description=f"Check {channel.mention}!",
-            colour=Colour.blue(),
-        )
-    elif ctx.message.guild:
-        player_id = ctx.message.author.id
-        member_: Member | None = ctx.message.guild.get_member(player_id)
-        if member_:
-            try:
-                await member_.send(
-                    embed=Embed(
-                        description=f"{output}",
-                        colour=Colour.blue(),
-                    ),
-                )
-            except Exception:
-                pass
-
-
 @bot.command(name="del")
 async def del_(ctx: Context, *args):
     """
@@ -1519,11 +1062,6 @@ async def del_(ctx: Context, *args):
     session.close()
 
 
-@bot.command()
-async def testleaderboard(ctx: Context):
-    await print_leaderboard(ctx.channel)
-
-
 # @bot.command()
 # @commands.check(is_admin)
 # async def imagetest(ctx: Context):
@@ -1548,75 +1086,6 @@ async def lt(ctx: Context):
     cropped.save(ntf.name)
     await ctx.message.channel.send(file=discord.File(ntf.name))
     ntf.close()
-
-
-@bot.tree.command(
-    name="gamehistory",
-    description="Privately displays your game history",
-)
-@discord.app_commands.guild_only()
-async def gamehistory(interaction: Interaction, count: int):
-    assert interaction.guild
-    if count > 10:
-        await interaction.response.send_message(
-            embed=Embed(
-                description="Count cannot exceed 10",
-                color=Colour.red(),
-            ),
-            ephemeral=True,
-        )
-        return
-    elif count < 1:
-        await interaction.response.send_message(
-            embed=Embed(
-                description="Count cannot be less than 1",
-                color=Colour.red(),
-            ),
-            ephemeral=True,
-        )
-        return
-
-    await interaction.response.defer(ephemeral=True, thinking=True)
-    session: SQLAlchemySession
-    with Session() as session:
-        finished_games: list[FinishedGame]
-        finished_games = (
-            session.query(FinishedGame)
-            .join(
-                FinishedGamePlayer,
-                FinishedGamePlayer.finished_game_id == FinishedGame.id,
-            )
-            .filter(FinishedGamePlayer.player_id == interaction.user.id)
-            .order_by(FinishedGame.finished_at.desc())
-            .limit(count)
-            .all()
-        )
-        if not finished_games:
-            await interaction.followup.send(
-                embed=Embed(
-                    description=f"{interaction.user.mention} has not played any games",
-                ),
-                ephemeral=True,
-                allowed_mentions=discord.AllowedMentions.none(),
-            )
-            return
-
-        embeds = []
-        finished_games.reverse()  # show most recent games last
-        for finished_game in finished_games:
-            # TODO: bold the callers name to make their name easier to see in the embed
-            embed: discord.Embed = create_finished_game_embed(
-                session, finished_game.id, interaction.guild.id
-            )
-            embed.timestamp = finished_game.finished_at
-            embeds.append(embed)
-
-        await interaction.followup.send(
-            content=f"Last {count} games for {interaction.user.mention}",
-            embeds=embeds,
-            ephemeral=True,
-            allowed_mentions=discord.AllowedMentions.none(),
-        )
 
 
 @bot.command()
@@ -1644,461 +1113,6 @@ async def pug(ctx: Context):
 #         embed_description=f"Delay between games set to {RE_ADD_DELAY}",
 #         colour=Colour.green(),
 #     )
-
-
-@bot.command()
-async def showgame(ctx: Context, game_id: str):
-    message = ctx.message
-    session = ctx.session
-    finished_game = (
-        session.query(FinishedGame)
-        .filter(FinishedGame.game_id.startswith(game_id))
-        .first()
-    )
-    if not finished_game:
-        await send_message(
-            message.channel,
-            embed_description=f"Could not find game: {game_id}",
-            colour=Colour.red(),
-        )
-        return
-
-    game_str = finished_game_str(finished_game)
-    await send_message(
-        message.channel,
-        embed_description=game_str,
-        colour=Colour.blue(),
-    )
-
-
-@bot.command()
-@commands.is_owner()
-async def showgamedebug(ctx: Context, game_id: str):
-    player_id = ctx.message.author.id
-
-    message = ctx.message
-    session = ctx.session
-    finished_game = (
-        session.query(FinishedGame)
-        .filter(FinishedGame.game_id.startswith(game_id))
-        .first()
-    )
-    if finished_game:
-        game_str = finished_game_str(finished_game, debug=True)
-        fgps: list[FinishedGamePlayer] = (
-            session.query(FinishedGamePlayer)
-            .filter(FinishedGamePlayer.finished_game_id == finished_game.id)
-            .all()
-        )
-        player_ids: list[int] = [fgp.player_id for fgp in fgps]
-        best_teams = get_n_best_finished_game_teams(
-            fgps, (len(fgps) + 1) // 2, finished_game.is_rated, 7
-        )
-        worst_teams = get_n_worst_finished_game_teams(
-            fgps, (len(fgps) + 1) // 2, finished_game.is_rated, 1
-        )
-        game_str += "\n**Most even team combinations:**"
-        for i, (_, best_team) in enumerate(best_teams):
-            # Every two pairings is the same
-            if i % 2 == 0:
-                continue
-            team0_players = best_team[: len(best_team) // 2]
-            team1_players = best_team[len(best_team) // 2 :]
-            game_str += f"\n{mock_finished_game_teams_str(team0_players, team1_players, finished_game.is_rated)}"
-        game_str += "\n\n**Least even team combination:**"
-        for _, worst_team in worst_teams:
-            team0_players = worst_team[: len(worst_team) // 2]
-            team1_players = worst_team[len(worst_team) // 2 :]
-            game_str += f"\n{mock_finished_game_teams_str(team0_players, team1_players, finished_game.is_rated)}"
-        # await send_message(
-        #     message.channel,
-        #     embed_description=game_str,
-        #     colour=Colour.blue(),
-        # )
-        if ctx.message.guild:
-            player_id = ctx.message.author.id
-            member_: Member | None = ctx.message.guild.get_member(player_id)
-            await send_message(
-                ctx.message.channel,
-                embed_description="Stats sent to DM",
-                colour=Colour.blue(),
-            )
-            if member_:
-                try:
-                    await member_.send(
-                        embed=Embed(
-                            description=game_str,
-                            colour=Colour.blue(),
-                        ),
-                    )
-                except Exception:
-                    pass
-
-    else:
-        in_progress_game: InProgressGame | None = (
-            session.query(InProgressGame)
-            .filter(InProgressGame.id.startswith(game_id))
-            .first()
-        )
-        if not in_progress_game:
-            await send_message(
-                message.channel,
-                embed_description=f"Could not find game: {game_id}",
-                colour=Colour.red(),
-            )
-            return
-        else:
-            game_str = in_progress_game_str(in_progress_game, debug=True)
-            queue: Queue = (
-                session.query(Queue)
-                .filter(Queue.id == in_progress_game.queue_id)
-                .first()
-            )
-            igps: list[InProgressGamePlayer] = (
-                session.query(InProgressGamePlayer)
-                .filter(InProgressGamePlayer.in_progress_game_id == in_progress_game.id)
-                .all()
-            )
-            player_ids: list[int] = [igp.player_id for igp in igps]
-            players: list[Player] = (
-                session.query(Player).filter(Player.id.in_(player_ids)).all()
-            )
-            best_teams = get_n_best_teams(
-                players, (len(players) + 1) // 2, queue.is_rated, 5
-            )
-            worst_teams = get_n_worst_teams(
-                players, (len(players) + 1) // 2, queue.is_rated, 1
-            )
-            game_str += "\n**Most even team combinations:**"
-            for _, best_team in best_teams:
-                team0_players = best_team[: len(best_team) // 2]
-                team1_players = best_team[len(best_team) // 2 :]
-                game_str += (
-                    f"\n{mock_teams_str(team0_players, team1_players, queue.is_rated)}"
-                )
-            game_str += "\n\n**Least even team combination:**"
-            for _, worst_team in worst_teams:
-                team0_players = worst_team[: len(worst_team) // 2]
-                team1_players = worst_team[len(worst_team) // 2 :]
-                game_str += (
-                    f"\n{mock_teams_str(team0_players, team1_players, queue.is_rated)}"
-                )
-            if ctx.message.guild:
-                player_id = ctx.message.author.id
-                member_: Member | None = ctx.message.guild.get_member(player_id)
-                await send_message(
-                    ctx.message.channel,
-                    embed_description="Stats sent to DM",
-                    colour=Colour.blue(),
-                )
-                if member_:
-                    try:
-                        await member_.send(
-                            embed=Embed(
-                                description=game_str,
-                                colour=Colour.blue(),
-                            ),
-                        )
-                    except Exception:
-                        pass
-
-            # await send_message(
-            #     message.channel,
-            #     embed_description=game_str,
-            #     colour=Colour.blue(),
-            # )
-
-
-@bot.tree.command(
-    name="mapstats",
-    description="For a given category, privately displays your winrate per map",
-)
-@discord.app_commands.rename(category_name="category")
-@discord.app_commands.describe(category_name="Optional name of a specific category")
-async def mapstats(interaction: Interaction, category_name: Optional[str] = None):
-    # TODO: merge with /stats by making this a subcommand
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        if category_name:
-            category: Category | None = (
-                session.query(Category).filter(Category.name == category_name).one()
-            )
-            # get the queues for each category, since rotations are per queue and not per category
-            queues: list[Queue] | None = (
-                session.query(Queue).filter(Queue.category_id == category.id).all()
-            )
-        else:
-            queues: list[Queue] | None = session.query(Queue).all()
-        if not queues:
-            await interaction.response.send_message(
-                embed=discord.Embed(
-                    description=f"Could not find any queues",
-                    color=discord.Color.red(),
-                ),
-                ephemeral=True,
-            )
-            return
-        rotation_ids = [queue.rotation_id for queue in queues]
-        maps: list[Map] | None = (
-            session.query(Map)
-            .join(RotationMap, RotationMap.map_id == Map.id)
-            .filter(RotationMap.rotation_id.in_(rotation_ids))
-            .order_by(Map.full_name)
-            .all()
-        )
-        if not maps:
-            await interaction.response.send_message(
-                embed=discord.Embed(
-                    description=f"Could not find any maps for category '{category_name}'",
-                    color=discord.Color.red(),
-                ),
-                ephemeral=True,
-            )
-            return
-        fgps: List[FinishedGamePlayer] | None = (
-            session.query(FinishedGamePlayer)
-            .filter(FinishedGamePlayer.player_id == interaction.user.id)
-            .all()
-        )
-        if not fgps:
-            await interaction.response.send_message(
-                embed=discord.Embed(
-                    description=f"You have not played any games",
-                    color=discord.Color.red(),
-                ),
-                ephemeral=True,
-            )
-            return
-        finished_game_ids: List[str] | None = [fgp.finished_game_id for fgp in fgps]
-        conditions = [FinishedGame.id.in_(finished_game_ids)]
-        if category_name:
-            conditions.append(FinishedGame.category_name == category_name)
-        fgs: List[FinishedGame] | None = (
-            session.query(FinishedGame).filter(*conditions).all()
-        )
-        if not fgs:
-            await interaction.response.send_message(
-                embed=discord.Embed(
-                    description=f"Could not find any finished games for you",
-                    color=discord.Color.red(),
-                ),
-                ephemeral=True,
-            )
-            return
-        fgps_by_finished_game_id: dict[str, FinishedGamePlayer] = {
-            fgp.finished_game_id: fgp for fgp in fgps
-        }
-        cols = []
-        for m in maps:
-            def map_stats(finished_games: list[FinishedGame]):
-                wins = [
-                    fg
-                    for fg in finished_games
-                    if fg.winning_team == fgps_by_finished_game_id[fg.id].team
-                ]
-                losses = [
-                    fg
-                    for fg in finished_games
-                    if fg.winning_team != fgps_by_finished_game_id[fg.id].team
-                    and fg.winning_team != -1
-                ]
-                ties = [fg for fg in finished_games if fg.winning_team == -1]
-                return len(wins), len(losses), len(ties)
-
-            fgs_for_map = [fg for fg in fgs if fg.map_full_name == m.full_name]
-            num_games = len(fgs_for_map)
-            if num_games <= 0:
-                continue
-            wins, losses, ties = map_stats(fgs_for_map)
-            wr = win_rate(wins, losses, ties)
-            cols.append(
-                [
-                    m.full_name,
-                    f"{wins}",
-                    f"{losses}",
-                    f"{ties}",
-                    num_games,
-                    f"{wr}%",
-                ]
-            )
-    table = table2ascii(
-        header=["Map", "W", "L", "T", "Total", "WR"],
-        body=cols,
-        style=PresetStyle.plain,
-        first_col_heading=True,
-        alignments=[
-            Alignment.LEFT,
-            Alignment.DECIMAL,
-            Alignment.DECIMAL,
-            Alignment.DECIMAL,
-            Alignment.DECIMAL,
-            Alignment.RIGHT,
-        ],
-    )
-    if category_name:
-        title = f"{interaction.user.display_name} Map Stats for {category_name}"
-    else:
-        title = f"{interaction.user.display_name} Overall Map Stats"
-    embed = discord.Embed(
-        title=title, description=code_block(table), color=discord.Color.blue()
-    )
-    await interaction.response.send_message(embed=embed, ephemeral=True)
-
-
-@bot.tree.command(
-    name="globalmapstats", description="Displays global statistics for each map"
-)
-@discord.app_commands.rename(category_name="category")
-@discord.app_commands.describe(category_name="Optional name of a specific category")
-async def globalmapstats(
-    interaction: discord.Interaction, category_name: Optional[str] = None
-):
-    # Explicitly does not use a discord.Embed, due to the limit of the Embed length (Note: this won't look pretty on mobile)
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        maps: list[Map] | None = session.query(Map).order_by(Map.full_name).all()
-
-        def map_stats(finished_games: list[FinishedGame]):
-            team0_wins = [fg for fg in finished_games if fg.winning_team == 0]
-            team1_wins = [fg for fg in finished_games if fg.winning_team == 1]
-            ties = [fg for fg in finished_games if fg.winning_team == -1]
-            return len(team0_wins), len(team1_wins), len(ties)
-
-        cols = []
-        for m in maps:
-            conditions = [FinishedGame.map_full_name == m.full_name]
-            if category_name:
-                conditions.append(FinishedGame.category_name == category_name)
-            finished_games = session.query(FinishedGame).filter(*conditions).all()
-            num_games = len(finished_games)
-            if num_games <= 0:
-                continue
-            team0_wins, team1_wins, ties = map_stats(finished_games)
-            team0_win_rate = win_rate(team0_wins, team1_wins, ties)
-            team1_win_rate = win_rate(team1_wins, team0_wins, ties)
-            cols.append(
-                [
-                    m.full_name,
-                    team0_wins,
-                    f"{team0_win_rate}%",
-                    team1_wins,
-                    f"{team1_win_rate}%",
-                    ties,
-                    len(finished_games),
-                ]
-            )
-    table = table2ascii(
-        header=["Map", "Team0", "WR", "Team1", "WR", "Ties", "Total"],
-        body=cols,
-        style=PresetStyle.plain,
-        first_col_heading=True,
-        alignments=[
-            Alignment.LEFT,
-            Alignment.DECIMAL,
-            Alignment.RIGHT,
-            Alignment.DECIMAL,
-            Alignment.RIGHT,
-            Alignment.DECIMAL,
-            Alignment.DECIMAL,
-        ],
-    )
-    if category_name:
-        content = f"Global Map Stats for **{category_name}**"
-    else:
-        content = "Global Map Stats"
-    content += f"\n{code_block(table)}"
-    await interaction.response.send_message(
-        content=content
-    )  # TODO: consider making this ephemeral, or giving the option to choose
-
-
-def win_rate(wins, losses, ties):
-    denominator = max(wins + losses + ties, 1)
-    return round(
-        100 * (wins + 0.5 * ties) / denominator, 1
-    )  # why are ties counted as half a win?
-
-
-@stats.autocomplete("category_name")
-@mapstats.autocomplete("category_name")
-async def category_name_autocomplete_with_user_id(
-    interaction: discord.Interaction, current: str
-):
-    # useful for when you want to filter the categories based on the ones the author has games played in
-    choices = []
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        result = (
-            session.query(Category.name, PlayerCategoryTrueskill.player_id)
-            .join(PlayerCategoryTrueskill)
-            .filter(PlayerCategoryTrueskill.player_id == interaction.user.id)
-            .order_by(Category.name)
-            .limit(25)  # discord only supports up to 25 choices
-            .all()
-        )
-        category_names: list[str] = [r[0] for r in result] if result else []
-        for name in category_names:
-            if current in name:
-                choices.append(
-                    discord.app_commands.Choice(
-                        name=name,
-                        value=name,
-                    )
-                )
-    return choices
-
-
-@globalmapstats.autocomplete("category_name")
-async def category_name_autocomplete_without_user_id(
-    interaction: discord.Interaction, current: str
-):
-    # useful for when you want all of the categories, regardless of whether the user has played games in them
-    choices = []
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        categories: list[Category] | None = (
-            session.query(Category)
-            .order_by(Category.name)
-            .limit(25)  # discord only supports up to 25 choices
-            .all()
-        )
-        if categories:
-            for category in categories:
-                if current in category.name:
-                    choices.append(
-                        discord.app_commands.Choice(
-                            name=category.name,
-                            value=category.name,
-                        )
-                    )
-    return choices
-
-
-def category_name_autocomplete(current: str, user_id: Optional[int] = None):
-    choices = []
-    session: sqlalchemy.orm.Session
-    with Session() as session:
-        conditions = []
-        if user_id:
-            conditions.append(PlayerCategoryTrueskill.player_id == user_id)
-        result = (
-            session.query(Category.name, PlayerCategoryTrueskill.player_id)
-            .join(PlayerCategoryTrueskill)
-            .filter(*conditions)
-            .order_by(Category.name)
-            .limit(25)  # discord only supports up to 25 choices
-            .all()
-        )
-        category_names: list[str] = [r[0] for r in result] if result else []
-        for name in category_names:
-            if current in name:
-                choices.append(
-                    discord.app_commands.Choice(
-                        name=name,
-                        value=name,
-                    )
-                )
-    return choices
 
 
 @bot.command()
